chore: move the command options into deploy-contract file
